feat: add autoFocus property to first field on each form
In some places, for example https://davidwalsh.name/react-autofocus, they say that _autoFocus_ is not enough.

However, in some places, for example https://stackoverflow.com/questions/42324423/autofocus-an-input-element-in-react-js#42324485
other said that _autoFocus_ should work.

By the moment, it is working, we can improve it in the future, for example scanning autoFocus fields in FormWithCaptcha or other places.